+++ pyforestscan_qgis/algorithms/advanced/__init__.py
@@ -3,10 +3,12 @@
 from .advanced_chm import AdvancedChmAlgorithm
 from .advanced_dtm import AdvancedDtmAlgorithm
 from .advanced_pad import AdvancedPadAlgorithm
+from .advanced_point_density import AdvancedPointDensityAlgorithm
 from .advanced_pai import AdvancedPaiAlgorithm
 from .advanced_canopy_cover import AdvancedCanopyCoverAlgorithm
 from .advanced_fhd import AdvancedFhdAlgorithm
 from .advanced_rumple import AdvancedRumpleAlgorithm
+from .advanced_voxel_stat import AdvancedVoxelStatAlgorithm
 from .normalize_hag import NormalizeHagAlgorithm
 from .point_cloud_preprocess import PointCloudPreprocessAlgorithm
 
@@ -14,10 +16,12 @@ __all__ = [
     "AdvancedChmAlgorithm",
     "AdvancedDtmAlgorithm",
     "AdvancedPadAlgorithm",
+    "AdvancedPointDensityAlgorithm",
     "AdvancedPaiAlgorithm",
     "AdvancedCanopyCoverAlgorithm",
     "AdvancedFhdAlgorithm",
     "AdvancedRumpleAlgorithm",
+    "AdvancedVoxelStatAlgorithm",
     "NormalizeHagAlgorithm",
     "PointCloudPreprocessAlgorithm",
 ]